refactor(agent-bridge): operations own themselves, MCP becomes a door (CHOO-1857/490)
The HTTP endpoint previously dispatched into the MCP tool registry, which
made the remote MCP server the foundation and HTTP its consumer. Since
Switch hosting MCP is the legacy path, that dependency ran the wrong way:
removing it would have meant rebuilding the operation layer.

Operations now live in their own registry that neither door owns. The MCP
server registers its tools from it and the HTTP endpoint dispatches into
it, so parity still cannot drift - but retiring a door is now deleting a
file rather than a refactor.

Operations take their arguments and nothing else: the FastMCP Context
parameter is gone from all 44, and the caller's identity and session key
come from a call context each door binds before dispatching. The operations
layer imports nothing from fastmcp, and mcp/server.py is 1700 lines
smaller.
