+++ core/tests/switch_core/bridges/agent/api/test_operations.py
@@ -1,8 +1,8 @@
 """HTTP operations front door (CHOO-1857 / CHOO-490).
 
-The property under test is parity: every MCP tool is reachable over HTTP under
-the same name, because both doors dispatch into one registry. If someone adds a
-tool and these fail, that is the point.
+The property under test is parity: every operation is reachable through both
+doors under the same name, because both are built from one registry. If someone
+adds an operation and these fail, that is the point.
 """
 
 from __future__ import annotations
@@ -16,7 +16,9 @@ from switch_core.bridges.agent.api.operations import (
     list_operations,
 )
 from switch_core.bridges.agent.mcp import server as mcp_server
-from switch_core.bridges.agent.mcp.callctx import (
+from switch_core.bridges.agent.operations import context as op_context
+from switch_core.bridges.agent.operations import get_operation
+from switch_core.bridges.agent.operations.callctx import (
     CallContext,
     current_call_context,
     reset_call_context,
@@ -30,24 +32,26 @@ async def _tool_names() -> set[str]:
     return {tool.name for tool in await mcp_server.mcp._list_tools()}
 
 
-async def test_every_mcp_tool_is_reachable_over_http() -> None:
-    assert set(await list_operations()) == await _tool_names()
+async def test_every_operation_is_registered_as_an_mcp_tool() -> None:
+    # The MCP door registers whatever the registry holds, so this is the
+    # parity guarantee: neither door can be missing an operation the other has.
+    assert set(list_operations()) == await _tool_names()
 
 
 async def test_operation_names_are_the_tool_names_verbatim() -> None:
     # One vocabulary: a translating runtime is POST /ops/${toolName}, nothing
     # more. Renaming or namespacing here would reintroduce a mapping table.
-    ops = await list_operations()
+    ops = list_operations()
     for expected in ("connect_to_room", "post_message", "assume_role", "list_tasks"):
         assert expected in ops
 
 
-async def test_operations_advertise_their_parameters() -> None:
-    ops = await list_operations()
-    params = ops["connect_to_room"]["parameters"]
-    assert "room_id" in params["properties"]
-    # `ctx` is a transport detail, not an argument an agent supplies.
-    assert "ctx" not in params["properties"]
+def test_operations_advertise_their_parameters() -> None:
+    params = list_operations()["connect_to_room"]["parameters"]
+    assert params["room_id"]["required"] is True
+    assert params["include_general_instructions"]["required"] is False
+    # Transport details are not arguments an agent supplies.
+    assert "ctx" not in params
 
 
 async def test_unknown_operation_names_what_is_available() -> None:
@@ -90,25 +94,16 @@ async def test_the_call_context_carries_agent_and_connection() -> None:
     """The operation sees who called and which connection it belongs to."""
     seen: dict[str, object] = {}
 
-    async def fake_op(room_id: str, ctx: object = None) -> str:
-        bound = current_call_context()
-        seen["agent_id"] = bound.agent_id if bound else None
-        seen["session_key"] = bound.session_key if bound else None
-        seen["ctx"] = ctx
+    async def fake_op(room_id: str) -> str:
+        seen["agent_id"] = op_context.get_agent_id()
+        seen["session_key"] = op_context.session_key()
         return "ok"
 
-    class _FakeTool:
-        name = "fake_op"
-        fn = staticmethod(fake_op)
-        description = "fake"
-        parameters: dict = {}
+    from switch_core.bridges.agent.operations import registry
 
-    original = mcp_server.mcp._list_tools
-
-    async def _list_tools():  # type: ignore[no-untyped-def]
-        return [*await original(), _FakeTool()]
-
-    mcp_server.mcp._list_tools = _list_tools  # type: ignore[method-assign]
+    registry._REGISTRY["fake_op"] = registry.Operation(
+        name="fake_op", fn=fake_op, description="fake"
+    )
     try:
         result = await call_operation(
             operation="fake_op",
@@ -117,14 +112,13 @@ async def test_the_call_context_carries_agent_and_connection() -> None:
             connection_id="conn-9",
         )
     finally:
-        mcp_server.mcp._list_tools = original  # type: ignore[method-assign]
+        registry._REGISTRY.pop("fake_op", None)
 
     assert result == "ok"
     assert seen["agent_id"] == AGENT
     # The connection is the caller's session key: this is what lets an
     # operation resolve the room binding without an MCP transport session.
     assert seen["session_key"] == "conn-9"
-    assert seen["ctx"] is None
 
 
 def test_the_call_context_is_cleared_after_the_call() -> None:
@@ -135,27 +129,34 @@ def test_the_call_context_is_cleared_after_the_call() -> None:
     assert current_call_context() is None
 
 
-async def test_agent_id_prefers_the_bound_context_over_the_request_scope() -> None:
+def test_agent_id_prefers_the_bound_context_over_the_request_scope() -> None:
     token = set_call_context(CallContext(agent_id="bound-agent", session_key=None))
     try:
         # No HTTP request in scope at all: resolving would fail if the bound
         # context were not consulted first.
-        assert mcp_server._get_agent_id() == "bound-agent"
+        assert op_context.get_agent_id() == "bound-agent"
     finally:
         reset_call_context(token)
 
 
-async def test_session_key_prefers_the_bound_context() -> None:
+def test_session_key_prefers_the_bound_context() -> None:
     token = set_call_context(CallContext(agent_id=AGENT, session_key="conn-7"))
     try:
-        assert mcp_server._session_key(None) == "conn-7"
+        assert op_context.session_key() == "conn-7"
     finally:
         reset_call_context(token)
 
 
-def test_session_key_falls_back_to_the_mcp_transport_session() -> None:
-    class _Ctx:
-        session_id = "mcp-session-3"
+def test_session_key_is_none_when_bound_to_nothing() -> None:
+    # Neither an HTTP call context nor an MCP session: operations that need a
+    # room report "not connected" rather than guessing one.
+    assert op_context.session_key() is None
 
-    assert mcp_server._session_key(_Ctx()) == "mcp-session-3"
-    assert mcp_server._session_key(None) is None
+
+def test_the_registry_refuses_duplicate_operation_names() -> None:
+    from switch_core.bridges.agent.operations import registry
+
+    existing = get_operation("post_message")
+    assert existing is not None
+    with pytest.raises(RuntimeError):
+        registry.operation(existing.fn)
